// src/redux/enums.tsx
export enum UartParity {
    noParity = "noParity",
    oddParity = "oddParity",
    evenParity = "evenParity"
}

export enum ConnectionUartStatus {
    notConnected = "notConnected",
    tryingToConnect = "tryingToConnect",
    connected = "connected",
    tryingToDisconnect = "tryingToDisconnect",
}


export enum FlowControlType {
    none = "none",
    hardware = "hardware"
}